DC-7696 Update models
