+++ src/main/scala/uk/gov/hmrc/common/message/model/Message.scala
@@ -16,14 +16,15 @@
 
 package uk.gov.hmrc.common.message.model
 
-import java.time.{ Instant, LocalDate }
-import org.mongodb.scala.bson.ObjectId
 import org.apache.commons.codec.binary.Base64
-import play.api.libs.json.{ Format, JsError, JsObject, JsResult, JsString, JsSuccess, JsValue, Json, OFormat, OWrites, Reads, Writes }
-import uk.gov.hmrc.domain.TaxIds._
+import org.mongodb.scala.bson.ObjectId
+import play.api.libs.json.*
+import uk.gov.hmrc.domain.TaxIds.*
 import uk.gov.hmrc.mongo.play.json.formats.MongoJavatimeFormats
 import uk.gov.hmrc.mongo.workitem.ProcessingStatus
 import uk.gov.hmrc.mongo.workitem.ProcessingStatus.ToDo
+
+import java.time.{ Instant, LocalDate }
 
 case class MessageContentParameters(data: ContentParameters, templateId: String)
 object MessageContentParameters {
@@ -130,7 +131,7 @@ case class Message(
 }
 
 object Message {
-  import MessageMongoFormats._
+  import MessageMongoFormats.*
 
   implicit val taxIdWithNameWrites: Writes[TaxIdWithName] = new Writes[TaxIdWithName] {
     override def writes(taxId: TaxIdWithName): JsValue =
@@ -266,4 +267,27 @@ object MailgunStatus {
     override def writes(status: MailgunStatus): JsValue = JsString(status.name)
   }
 
+}
+
+enum LifecycleStatusType(val entryName: String) {
+  case Submitted extends LifecycleStatusType("SUBMITTED")
+  case SubmissionFailed extends LifecycleStatusType("SUBMISSION_FAILED")
+  case Delivered extends LifecycleStatusType("DELIVERED")
+  case DeliveryFailed extends LifecycleStatusType("DELIVERY_FAILED")
+  case Responded extends LifecycleStatusType("RESPONDED")
+}
+
+object LifecycleStatusType {
+  implicit val format: Format[LifecycleStatusType] = new Format[LifecycleStatusType] {
+    def reads(json: JsValue): JsResult[LifecycleStatusType] = json match {
+      case JsString("SUBMITTED")         => JsSuccess(Submitted)
+      case JsString("SUBMISSION_FAILED") => JsSuccess(SubmissionFailed)
+      case JsString("DELIVERED")         => JsSuccess(Delivered)
+      case JsString("DELIVERY_FAILED")   => JsSuccess(DeliveryFailed)
+      case JsString("RESPONDED")         => JsSuccess(Responded)
+      case _                             => JsError("Invalid Lifecycle Status Type")
+    }
+
+    def writes(status: LifecycleStatusType): JsValue = JsString(status.entryName)
+  }
 }
